Use the taxon distribution data from the APC to determine state level native and introduced origin status

This function processes the geographic data available in the APC and returns state level native, introduced and more complicated origins status for all taxa.

Usage

create_species_state_origin_matrix(resources = load_taxonomic_resources())

Arguments

resources

the taxonomic resources required to make the summary statistics. Loading this can be slow, so call load_taxonomic_resources separately to greatly speed this function up and pass the resources in.

Value

A tibble with columns representing each state and rows representing each species. The values in each cell represent the origin of the species in that state.

See also

Examples

create_species_state_origin_matrix()
#> Loading resources...
#> ...done
#> # A tibble: 26,229 × 19
#>    species     NT    Qld   WA    ChI   NSW   SA    Vic   Tas   ACT   NI    LHI  
#>    <chr>       <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 Cycas angu… nati… nati… not … not … not … not … not … not … not … not … not …
#>  2 Cycas aren… nati… not … not … not … not … not … not … not … not … not … not …
#>  3 Cycas arms… nati… not … not … not … not … not … not … not … not … not … not …
#>  4 Cycas arnh… nati… not … not … not … not … not … not … not … not … not … not …
#>  5 Cycas bade… not … nati… not … not … not … not … not … not … not … not … not …
#>  6 Cycas basa… not … not … nati… not … not … not … not … not … not … not … not …
#>  7 Cycas brun… nati… nati… not … not … not … not … not … not … not … not … not …
#>  8 Cycas cair… not … nati… not … not … not … not … not … not … not … not … not …
#>  9 Cycas calc… nati… not … not … not … not … not … not … not … not … not … not …
#> 10 Cycas cana… nati… not … not … not … not … not … not … not … not … not … not …
#> # ℹ 26,219 more rows
#> # ℹ 7 more variables: MI <chr>, HI <chr>, MDI <chr>, CoI <chr>, CSI <chr>,
#> #   AR <chr>, CaI <chr>