Find taxonomic alignments for a list of names to a version of the Australian Plant Census (APC) through standardizing formatting and checking for spelling issues

This function uses Australian Plant Census (APC) & the Australian Plant Name Index (APNI) to find taxonomic alignments for a list of names.

Usage

align_taxa(
  original_name,
  output = NULL,
  full = FALSE,
  resources = load_taxonomic_resources(),
  fuzzy_abs_dist = 3,
  fuzzy_rel_dist = 0.2,
  fuzzy_matches = TRUE,
  imprecise_fuzzy_matches = FALSE,
  APNI_matches = TRUE,
  identifier = NA_character_
)

Arguments

original_name

A list of names to query for taxonomic alignments.

output

(optional) The name of the file to save the results to.

full

Parameter to determine how many columns are output

resources

the taxonomic resources used to align the taxa names. Loading this can be slow, so call load_taxonomic_resources separately to greatly speed this function up and pass the resources in.

fuzzy_abs_dist

The number of characters allowed to be different for a fuzzy match.

fuzzy_rel_dist

The proportion of characters allowed to be different for a fuzzy match.

fuzzy_matches

Fuzzy matches are turned on as a default. The relative and absolute distances allowed for fuzzy matches to species and infraspecific taxon names are defined by the parameters fuzzy_abs_dist and fuzzy_rel_dist

imprecise_fuzzy_matches

Imprecise fuzzy matches are turned off as a default.

APNI_matches

Name matches to the APNI (Australian Plant Names Index) are turned off as a default.

identifier

A dataset, location or other identifier, which defaults to NA.

Value

A tibble with columns: original_name, cleaned_name, aligned_name, taxonomic_dataset, known, and checked.

See also

Examples

align_taxa(c("Poa annua", "Abies alba"))
#> Checking alignments of 2 taxa
#>   -> 0 names already matched; 0 names checked but without a match; 2 taxa yet to be checked
#> Loading resources...
#> ...done
#> # A tibble: 2 × 7
#>   original_name cleaned_name aligned_name taxonomic_dataset taxon_rank
#>   <chr>         <chr>        <chr>        <chr>             <chr>     
#> 1 Poa annua     Poa annua    Poa annua    APC               species   
#> 2 Abies alba    Abies alba   NA           NA                NA        
#> # ℹ 2 more variables: aligned_reason <chr>, alignment_code <chr>