State level native and introduced origin status

This function uses the taxon distribution data from the APC to determine state level native and introduced origin status.

This function processes the geographic data available in the APC and returns state level native, introduced and more complicated origins status for all taxa.

Usage

create_species_state_origin_matrix(resources = load_taxonomic_resources())

Arguments

resources

the taxonomic resources required to make the summary statistics. Loading this can be slow, so call load_taxonomic_resources separately to greatly speed this function up and pass the resources in.

Value

A tibble with columns representing each state and rows representing each species. The values in each cell represent the origin of the species in that state.

See also

Examples

create_species_state_origin_matrix()
#> 
#> Loading resources into memory...
#> ================================================================================================================================================================
#> ...done
#> # A tibble: 26,509 × 19
#>    species     WA    NT    Qld   SA    NSW   Vic   Tas   ChI   LHI   CoI   NI   
#>    <chr>       <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 Acanthus e… nati… nati… nati… not … not … not … not … not … not … not … not …
#>  2 Acanthus i… not … nati… nati… not … not … not … not … not … not … not … not …
#>  3 Acanthus m… natu… not … doub… natu… natu… natu… natu… not … not … not … not …
#>  4 Andrograph… not … natu… doub… not … not … not … not … nati… not … not … not …
#>  5 Asystasia … not … not … not … not … not … not … not … nati… not … not … not …
#>  6 Asystasia … not … not … nati… not … not … not … not … not … not … not … not …
#>  7 Asystasia … natu… natu… natu… not … natu… not … not … natu… not … not … not …
#>  8 Asystasia … not … not … not … not … not … not … not … natu… not … not … not …
#>  9 Asystasia … not … not … nati… not … not … not … not … not … not … not … not …
#> 10 Asystasia … not … not … not … not … not … not … not … natu… not … not … not …
#> # ℹ 26,499 more rows
#> # ℹ 7 more variables: ACT <chr>, CSI <chr>, AR <chr>, MI <chr>, HI <chr>,
#> #   MDI <chr>, CaI <chr>