Create a lookup table to help fix the taxonomy for a list of Australian plant species

This function takes a list of Australian plant species that needs to be reconciled with current taxonomy and generates a lookup table to help fix the taxonomy. The lookup table contains the original species names, the aligned species names, and additional taxonomic information such as taxon IDs and genera.

Usage

create_taxonomic_update_lookup(
  taxa,
  stable_or_current_data = "stable",
  version = default_version(),
  one_to_many = "return_all",
  full = FALSE,
  resources = load_taxonomic_resources(stable_or_current_data = stable_or_current_data,
    version = version),
  output = NULL
)

Arguments

taxa

A list of Australian plant species that needs to be reconciled with current taxonomy.

stable_or_current_data

either "stable" for a consistent version, or "current" for the leading edge version.

version

The version number of the dataset to use.

one_to_many

How to handle one_to_many taxonomic matches. Default is "return_all". The other options are "collapse_to_higher_taxon" and "most_likely_species". most_likely_species defaults to the original_name if that name is accepted by the APC; this will be right for certain species subsets, but make errors in other cases, use with caution.

full

logical for whether the full lookup table is returned or just the two key columns

resources

These are the taxonomic resources used for cleaning, this will default to loading them from a local place on your computer. If this is to be called repeatedly, it's much faster to load the resources using load_taxonomic_resources separately and pass the data in.

output

file path to save the intermediate output to

Value

A lookup table containing the original species names, the aligned species names, and additional taxonomic information such as taxon IDs and genera.

See also

Examples

resources <- load_taxonomic_resources()
#> Loading resources...
#> ...done
create_taxonomic_update_lookup(c("Eucalyptus regnans",
                                 "Acacia melanoxylon",
                                 "Banksia integrifolia",
                                 "Not a species"),
                                 resources=resources)
#> Checking alignments of 4 taxa
#>   -> 0 names already matched; 0 names checked but without a match; 4 taxa yet to be checked
#> # A tibble: 3 × 5
#>   original_name      aligned_name apc_name aligned_reason taxonomic_status_of_…¹
#>   <chr>              <chr>        <chr>    <chr>          <chr>                 
#> 1 Eucalyptus regnans Eucalyptus … Eucalyp… match_06. Aut… accepted              
#> 2 Acacia melanoxylon Acacia mela… Acacia … match_06. Aut… accepted              
#> 3 Banksia integrifo… Banksia int… Banksia… match_06. Aut… accepted              
#> # ℹ abbreviated name: ¹​taxonomic_status_of_aligned_name