Load taxonomic resources from either stable or current versions of APC and APNI

Loads taxonomic resources into the global environment. This function accesses taxonomic data from a dataset using the provided version number or the default version. The loaded data contains two lists: APC and APNI, which contain taxonomic information about plant species in Australia. The function creates several data frames by filtering and selecting data from the loaded lists.

Usage

load_taxonomic_resources(
  stable_or_current_data = "stable",
  version = default_version(),
  reload = FALSE
)

Arguments

stable_or_current_data

Type of dataset to access. The default is "stable", which loads the dataset from a github archived file. If set to "current", the dataset will be loaded from a URL which is the cutting edge version, but this may change at any time without notice.

version

The version number of the dataset to use. Defaults to the default version.

reload

A logical indicating whether to reload the dataset from the data source. Defaults to FALSE.

Value

The taxonomic resources data loaded into the global environment.

Examples

load_taxonomic_resources(stable_or_current_data="stable",version="0.0.2.9000")
#> Loading resources...
#> ...done
#> $APC
#> # A tibble: 109,624 × 37
#>    taxonID    nameType acceptedNameUsageID acceptedNameUsage nomenclaturalStatus
#>  * <chr>      <chr>    <chr>               <chr>             <chr>              
#>  1 https://i… scienti… https://id.biodive… Plantae Haeckel   NA                 
#>  2 https://i… scienti… https://id.biodive… Charophyta Sachs  NA                 
#>  3 https://i… scienti… https://id.biodive… Charophyta Sachs  orth. var.         
#>  4 https://i… scienti… https://id.biodive… Equisetopsida C.… NA                 
#>  5 https://i… scienti… https://id.biodive… Equisetopsida C.… orth. var.         
#>  6 https://i… scienti… https://id.biodive… Equisetopsida C.… NA                 
#>  7 https://i… scienti… https://id.biodive… Cycadidae Pax     NA                 
#>  8 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#>  9 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#> 10 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#> # ℹ 109,614 more rows
#> # ℹ 32 more variables: nomIlleg <chr>, nomInval <chr>, taxonomicStatus <chr>,
#> #   proParte <lgl>, scientificName <chr>, scientificNameID <chr>,
#> #   canonicalName <chr>, scientificNameAuthorship <chr>,
#> #   parentNameUsageID <chr>, taxonRank <chr>, taxonRankSortOrder <dbl>,
#> #   kingdom <chr>, class <chr>, subclass <chr>, family <chr>,
#> #   taxonConceptID <chr>, nameAccordingTo <chr>, nameAccordingToID <chr>, …
#> 
#> $APNI
#> # A tibble: 132,385 × 47
#>    scientificNameID     nameType scientificName scientificNameHTML canonicalName
#>  * <chr>                <chr>    <chr>          <chr>              <chr>        
#>  1 https://id.biodiver… scienti… Acanthus ebra… <scientific><name… Acanthus ebr…
#>  2 https://id.biodiver… autonym  Acanthus ebra… <scientific><name… Acanthus ebr…
#>  3 https://id.biodiver… scienti… Acanthus ebra… <scientific><name… Acanthus ebr…
#>  4 https://id.biodiver… autonym  Acanthus ilic… <scientific><name… Acanthus ili…
#>  5 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  6 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  7 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  8 https://id.biodiver… scienti… Acanthus moll… <scientific><name… Acanthus mol…
#>  9 https://id.biodiver… scienti… Acanthus volu… <scientific><name… Acanthus vol…
#> 10 https://id.biodiver… scienti… Acanthus L.    <scientific><name… Acanthus     
#> # ℹ 132,375 more rows
#> # ℹ 42 more variables: canonicalNameHTML <chr>, nameElement <chr>,
#> #   nomenclaturalStatus <chr>, scientificNameAuthorship <chr>, autonym <lgl>,
#> #   hybrid <lgl>, cultivar <lgl>, formula <lgl>, scientific <lgl>,
#> #   nomInval <lgl>, nomIlleg <lgl>, namePublishedIn <chr>,
#> #   namePublishedInID <chr>, namePublishedInYear <dbl>, nameInstanceType <chr>,
#> #   nameAccordingToID <chr>, nameAccordingTo <chr>, originalNameUsage <chr>, …
#> 
#> $`APC list (accepted)`
#> # A tibble: 30,538 × 13
#>    canonicalName     scientificName     taxonomicStatus ID    nameType taxonRank
#>    <chr>             <chr>              <chr>           <chr> <chr>    <chr>    
#>  1 Cycas angulata    Cycas angulata R.… accepted        http… scienti… Species  
#>  2 Cycas arenicola   Cycas arenicola K… accepted        http… scienti… Species  
#>  3 Cycas armstrongii Cycas armstrongii… accepted        http… scienti… Species  
#>  4 Cycas arnhemica   Cycas arnhemica K… accepted        http… scienti… Species  
#>  5 Cycas badensis    Cycas badensis K.… accepted        http… scienti… Species  
#>  6 Cycas basaltica   Cycas basaltica C… accepted        http… scienti… Species  
#>  7 Cycas brunnea     Cycas brunnea K.D… accepted        http… scienti… Species  
#>  8 Cycas cairnsiana  Cycas cairnsiana … accepted        http… scienti… Species  
#>  9 Cycas calcicola   Cycas calcicola M… accepted        http… scienti… Species  
#> 10 Cycas canalis     Cycas canalis K.D… accepted        http… scienti… Species  
#> # ℹ 30,528 more rows
#> # ℹ 7 more variables: stripped_canonical <chr>, stripped_canonical2 <chr>,
#> #   stripped_scientific <chr>, binomial <chr>, genus <chr>, trinomial <chr>,
#> #   taxonomic_ref <chr>
#> 
#> $`APC list (known names)`
#> # A tibble: 69,837 × 13
#>    canonicalName         scientificName taxonomicStatus ID    nameType taxonRank
#>    <chr>                 <chr>          <chr>           <chr> <chr>    <chr>    
#>  1 Cycas brunnea         Cycas brunnea… pro parte misa… http… scienti… Species  
#>  2 Cycas media f. inerm… Cycas media f… nomenclatural … http… scienti… Forma    
#>  3 Cycas media var. ine… Cycas media v… nomenclatural … http… scienti… Varietas 
#>  4 Cycas arnhemica subs… Cycas arnhemi… taxonomic syno… http… scienti… Subspeci…
#>  5 Cycas arnhemica subs… Cycas arnhemi… nomenclatural … http… autonym  Subspeci…
#>  6 Cycas arnhemica subs… Cycas arnhemi… taxonomic syno… http… scienti… Subspeci…
#>  7 Cycas media var. bas… Cycas media v… nomenclatural … http… scienti… Varietas 
#>  8 Cycas canalis subsp.… Cycas canalis… nomenclatural … http… autonym  Subspeci…
#>  9 Cycas canalis subsp.… Cycas canalis… taxonomic syno… http… scienti… Subspeci…
#> 10 Cycas media var. fur… Cycas media v… nomenclatural … http… scienti… Varietas 
#> # ℹ 69,827 more rows
#> # ℹ 7 more variables: stripped_canonical <chr>, stripped_canonical2 <chr>,
#> #   stripped_scientific <chr>, binomial <chr>, genus <chr>, trinomial <chr>,
#> #   taxonomic_ref <chr>
#> 
#> $`APNI names`
#> # A tibble: 15,377 × 13
#>    canonicalName         scientificName ID    nameType taxonRank taxonomicStatus
#>    <chr>                 <chr>          <chr> <chr>    <chr>     <chr>          
#>  1 (Eucalyptus globulus… (Eucalyptus g… http… hybrid-… Species   unplaced for A…
#>  2 (Eucalyptus globulus… (Eucalyptus g… http… hybrid-… Species   unplaced for A…
#>  3 (Eucalyptus johnston… (Eucalyptus j… http… hybrid-… Species   unplaced for A…
#>  4 (Eucalyptus johnston… (Eucalyptus j… http… hybrid-… Species   unplaced for A…
#>  5 (Eucalyptus urnigera… (Eucalyptus u… http… hybrid-… Species   unplaced for A…
#>  6 (Eucalyptus urnigera… (Eucalyptus u… http… hybrid-… Species   unplaced for A…
#>  7 Abacopteris presliana Abacopteris p… http… scienti… Species   unplaced for A…
#>  8 Abarema clypearia f.… Abarema clype… http… scienti… Forma     unplaced for A…
#>  9 Abarema pruinosa      Abarema pruin… http… scienti… Species   unplaced for A…
#> 10 Abelia rupestris      Abelia rupest… http… scienti… Species   unplaced for A…
#> # ℹ 15,367 more rows
#> # ℹ 7 more variables: stripped_canonical <chr>, stripped_canonical2 <chr>,
#> #   stripped_scientific <chr>, binomial <chr>, trinomial <chr>, genus <chr>,
#> #   taxonomic_ref <chr>
#> 
#> $genera_accepted
#> # A tibble: 3,094 × 8
#>    canonicalName acceptedNameUsage scientificName taxonomicStatus ID    nameType
#>    <chr>         <chr>             <chr>          <chr>           <chr> <chr>   
#>  1 Cycas         Cycas L.          Cycas L.       accepted        http… scienti…
#>  2 Bowenia       Bowenia Hook.f.   Bowenia Hook.… accepted        http… scienti…
#>  3 Lepidozamia   Lepidozamia Regel Lepidozamia R… accepted        http… scienti…
#>  4 Macrozamia    Macrozamia Miq.   Macrozamia Mi… accepted        http… scienti…
#>  5 Equisetum     Equisetum L.      Equisetum L.   accepted        http… scienti…
#>  6 Isoetes       Isoetes L.        Isoetes L.     accepted        http… scienti…
#>  7 Huperzia      Huperzia Bernh.   Huperzia Bern… accepted        http… scienti…
#>  8 Lycopodiella  Lycopodiella Hol… Lycopodiella … accepted        http… scienti…
#>  9 Lycopodium    Lycopodium L.     Lycopodium L.  accepted        http… scienti…
#> 10 Phlegmariurus Phlegmariurus Ho… Phlegmariurus… accepted        http… scienti…
#> # ℹ 3,084 more rows
#> # ℹ 2 more variables: taxonRank <chr>, taxonomic_ref <chr>
#> 
#> $genera_known
#> # A tibble: 3,028 × 8
#>    canonicalName acceptedNameUsage scientificName taxonomicStatus ID    nameType
#>    <chr>         <chr>             <chr>          <chr>           <chr> <chr>   
#>  1 Encephallart… Encephalartos Le… Encephallarte… orthographic v… http… scienti…
#>  2 Encephalartus Encephalartos Le… Encephalartus… orthographic v… http… scienti…
#>  3 Encephalartos Encephalartos Le… Encephalartos… excluded        http… scienti…
#>  4 Catakidozamia Lepidozamia Regel Catakidozamia… taxonomic syno… http… scienti…
#>  5 Zamia         Zamia L.          Zamia L.       excluded        http… scienti…
#>  6 Calamaria     Isoetes L.        Calamaria Boe… taxonomic syno… http… scienti…
#>  7 Lateristachys Lycopodiella Hol… Lateristachys… taxonomic syno… http… scienti…
#>  8 Lepidotis     Lycopodiella Hol… Lepidotis P.B… taxonomic syno… http… scienti…
#>  9 Austrolycopo… Lycopodium L.     Austrolycopod… taxonomic syno… http… scienti…
#> 10 Diphasium     Lycopodium L.     Diphasium C.P… taxonomic syno… http… scienti…
#> # ℹ 3,018 more rows
#> # ℹ 2 more variables: taxonRank <chr>, taxonomic_ref <chr>
#> 
#> $genera_APNI
#> # A tibble: 1,445 × 6
#>    canonicalName taxonomicStatus nameType taxonRank scientificName taxonomic_ref
#>    <chr>         <chr>           <chr>    <chr>     <chr>          <chr>        
#>  1 Beloperone    unplaced        scienti… Genus     Beloperone Ne… APNI         
#>  2 Hemiadelphis  unplaced        scienti… Genus     Hemiadelphis … APNI         
#>  3 Leptosiphoni… unplaced        scienti… Genus     Leptosiphoniu… APNI         
#>  4 Megaskepasma  unplaced        scienti… Genus     Megaskepasma … APNI         
#>  5 Meyenia       unplaced        scienti… Genus     Meyenia Nees   APNI         
#>  6 Simonachne    unplaced        scienti… Genus     Simonachne E.… APNI         
#>  7 Thyrsacanthus unplaced        scienti… Genus     Thyrsacanthus… APNI         
#>  8 Synarrhena    unplaced        scienti… Genus     Synarrhena F.… APNI         
#>  9 Allosorus     unplaced        scienti… Genus     Allosorus Ber… APNI         
#> 10 Austrogramme  unplaced        scienti… Genus     Austrogramme … APNI         
#> # ℹ 1,435 more rows
#> 
#> $genera_all
#> # A tibble: 7,513 × 9
#>    canonicalName acceptedNameUsage scientificName taxonomicStatus ID    nameType
#>    <chr>         <chr>             <chr>          <chr>           <chr> <chr>   
#>  1 Cycas         Cycas L.          Cycas L.       accepted        http… scienti…
#>  2 Bowenia       Bowenia Hook.f.   Bowenia Hook.… accepted        http… scienti…
#>  3 Lepidozamia   Lepidozamia Regel Lepidozamia R… accepted        http… scienti…
#>  4 Macrozamia    Macrozamia Miq.   Macrozamia Mi… accepted        http… scienti…
#>  5 Equisetum     Equisetum L.      Equisetum L.   accepted        http… scienti…
#>  6 Isoetes       Isoetes L.        Isoetes L.     accepted        http… scienti…
#>  7 Huperzia      Huperzia Bernh.   Huperzia Bern… accepted        http… scienti…
#>  8 Lycopodiella  Lycopodiella Hol… Lycopodiella … accepted        http… scienti…
#>  9 Lycopodium    Lycopodium L.     Lycopodium L.  accepted        http… scienti…
#> 10 Phlegmariurus Phlegmariurus Ho… Phlegmariurus… accepted        http… scienti…
#> # ℹ 7,503 more rows
#> # ℹ 3 more variables: taxonRank <chr>, taxonomic_ref <chr>, cleaned_name <chr>
#> 
#> $family_accepted
#> # A tibble: 307 × 37
#>    taxonID    nameType acceptedNameUsageID acceptedNameUsage nomenclaturalStatus
#>    <chr>      <chr>    <chr>               <chr>             <chr>              
#>  1 https://i… scienti… https://id.biodive… Cycadaceae Pers.  nom. cons.         
#>  2 https://i… scienti… https://id.biodive… Zamiaceae Horan.  nom. cons.         
#>  3 https://i… scienti… https://id.biodive… Equisetaceae Mic… NA                 
#>  4 https://i… scienti… https://id.biodive… Isoetaceae Rchb.  NA                 
#>  5 https://i… scienti… https://id.biodive… Lycopodiaceae P.… NA                 
#>  6 https://i… scienti… https://id.biodive… Selaginellaceae … NA                 
#>  7 https://i… scienti… https://id.biodive… Apiaceae Lindl.   nom. cons.         
#>  8 https://i… scienti… https://id.biodive… Araliaceae Juss.  nom. cons.         
#>  9 https://i… scienti… https://id.biodive… Griseliniaceae J… NA                 
#> 10 https://i… scienti… https://id.biodive… Myodocarpaceae D… NA                 
#> # ℹ 297 more rows
#> # ℹ 32 more variables: nomIlleg <chr>, nomInval <chr>, taxonomicStatus <chr>,
#> #   proParte <lgl>, scientificName <chr>, scientificNameID <chr>,
#> #   canonicalName <chr>, scientificNameAuthorship <chr>,
#> #   parentNameUsageID <chr>, taxonRank <chr>, taxonRankSortOrder <dbl>,
#> #   kingdom <chr>, class <chr>, subclass <chr>, family <chr>,
#> #   taxonConceptID <chr>, nameAccordingTo <chr>, nameAccordingToID <chr>, …
#>