Load taxonomic resources from either stable or current versions of APC and APNI

Loads taxonomic resources into the global environment. This function accesses taxonomic data from a dataset using the provided version number or the default version. The loaded data contains two lists: APC and APNI, which contain taxonomic information about plant species in Australia. The function creates several data frames by filtering and selecting data from the loaded lists.

Usage

load_taxonomic_resources(
  stable_or_current_data = "stable",
  version = default_version(),
  reload = FALSE
)

Arguments

stable_or_current_data

Type of dataset to access. The default is "stable", which loads the dataset from a github archived file. If set to "current", the dataset will be loaded from a URL which is the cutting edge version, but this may change at any time without notice.

version

The version number of the dataset to use. Defaults to the default version.

reload

A logical indicating whether to reload the dataset from the data source. Defaults to FALSE.

Value

The taxonomic resources data loaded into the global environment.

Examples

load_taxonomic_resources(stable_or_current_data="stable",version="0.0.2.9000")
#> Loading resources...
#> File downloaded successfully.
#> File downloaded successfully.
#> ...done
#> $APC
#> # A tibble: 109,624 × 38
#>    taxon_ID                 name_type accepted_name_usage_ID accepted_name_usage
#>    <chr>                    <chr>     <chr>                  <chr>              
#>  1 https://id.biodiversity… scientif… https://id.biodiversi… Plantae Haeckel    
#>  2 https://id.biodiversity… scientif… https://id.biodiversi… Charophyta Sachs   
#>  3 https://id.biodiversity… scientif… https://id.biodiversi… Charophyta Sachs   
#>  4 https://id.biodiversity… scientif… https://id.biodiversi… Equisetopsida C.Ag…
#>  5 https://id.biodiversity… scientif… https://id.biodiversi… Equisetopsida C.Ag…
#>  6 https://id.biodiversity… scientif… https://id.biodiversi… Equisetopsida C.Ag…
#>  7 https://id.biodiversity… scientif… https://id.biodiversi… Cycadidae Pax      
#>  8 https://id.biodiversity… scientif… https://id.biodiversi… Cycadales Pers. ex…
#>  9 https://id.biodiversity… scientif… https://id.biodiversi… Cycadales Pers. ex…
#> 10 https://id.biodiversity… scientif… https://id.biodiversi… Cycadales Pers. ex…
#> # ℹ 109,614 more rows
#> # ℹ 34 more variables: nomenclaturalStatus <chr>, nomIlleg <chr>,
#> #   nomInval <chr>, taxonomic_status <chr>, pro_parte <lgl>,
#> #   scientific_name <chr>, scientific_name_ID <chr>, canonical_name <chr>,
#> #   scientific_name_authorship <chr>, parentNameUsageID <chr>,
#> #   taxon_rank <chr>, taxon_rank_sort_order <dbl>, kingdom <chr>, class <chr>,
#> #   subclass <chr>, family <chr>, taxonConceptID <chr>, …
#> 
#> $APNI
#> # A tibble: 132,385 × 48
#>    scientific_name_ID               name_type scientific_name scientificNameHTML
#>    <chr>                            <chr>     <chr>           <chr>             
#>  1 https://id.biodiversity.org.au/… scientif… Acanthus ebrac… <scientific><name…
#>  2 https://id.biodiversity.org.au/… autonym   Acanthus ebrac… <scientific><name…
#>  3 https://id.biodiversity.org.au/… scientif… Acanthus ebrac… <scientific><name…
#>  4 https://id.biodiversity.org.au/… autonym   Acanthus ilici… <scientific><name…
#>  5 https://id.biodiversity.org.au/… scientif… Acanthus ilici… <scientific><name…
#>  6 https://id.biodiversity.org.au/… scientif… Acanthus ilici… <scientific><name…
#>  7 https://id.biodiversity.org.au/… scientif… Acanthus ilici… <scientific><name…
#>  8 https://id.biodiversity.org.au/… scientif… Acanthus molli… <scientific><name…
#>  9 https://id.biodiversity.org.au/… scientif… Acanthus volub… <scientific><name…
#> 10 https://id.biodiversity.org.au/… scientif… Acanthus L.     <scientific><name…
#> # ℹ 132,375 more rows
#> # ℹ 44 more variables: canonical_name <chr>, canonicalNameHTML <chr>,
#> #   name_element <chr>, nomenclaturalStatus <chr>,
#> #   scientific_name_authorship <chr>, autonym <lgl>, hybrid <lgl>,
#> #   cultivar <lgl>, formula <lgl>, scientific <lgl>, nomInval <lgl>,
#> #   nomIlleg <lgl>, namePublishedIn <chr>, namePublishedInID <chr>,
#> #   namePublishedInYear <dbl>, nameInstanceType <chr>, …
#> 
#> $`APC list (accepted)`
#> # A tibble: 30,538 × 15
#>    canonical_name   scientific_name taxonomic_status taxon_ID scientific_name_ID
#>    <chr>            <chr>           <chr>            <chr>    <chr>             
#>  1 Cycas angulata   Cycas angulata… accepted         https:/… https://id.biodiv…
#>  2 Cycas arenicola  Cycas arenicol… accepted         https:/… https://id.biodiv…
#>  3 Cycas armstrong… Cycas armstron… accepted         https:/… https://id.biodiv…
#>  4 Cycas arnhemica  Cycas arnhemic… accepted         https:/… https://id.biodiv…
#>  5 Cycas badensis   Cycas badensis… accepted         https:/… https://id.biodiv…
#>  6 Cycas basaltica  Cycas basaltic… accepted         https:/… https://id.biodiv…
#>  7 Cycas brunnea    Cycas brunnea … accepted         https:/… https://id.biodiv…
#>  8 Cycas cairnsiana Cycas cairnsia… accepted         https:/… https://id.biodiv…
#>  9 Cycas calcicola  Cycas calcicol… accepted         https:/… https://id.biodiv…
#> 10 Cycas canalis    Cycas canalis … accepted         https:/… https://id.biodiv…
#> # ℹ 30,528 more rows
#> # ℹ 10 more variables: accepted_name_usage_ID <chr>, name_type <chr>,
#> #   taxon_rank <chr>, genus <chr>, stripped_canonical <chr>,
#> #   stripped_canonical2 <chr>, stripped_scientific <chr>, binomial <chr>,
#> #   trinomial <chr>, taxonomic_dataset <chr>
#> 
#> $`APC list (known names)`
#> # A tibble: 69,647 × 15
#>    canonical_name   scientific_name taxonomic_status taxon_ID scientific_name_ID
#>    <chr>            <chr>           <chr>            <chr>    <chr>             
#>  1 Pterostylis cli… Pterostylis cl… alternative name https:/… https://id.biodiv…
#>  2 Pterostylis cre… Pterostylis cr… alternative name https:/… https://id.biodiv…
#>  3 Eucalyptus resi… Eucalyptus res… alternative name https:/… https://id.biodiv…
#>  4 Macrozamia paul… Macrozamia pau… basionym         https:/… https://id.biodiv…
#>  5 Lycopodium creb… Lycopodium cre… basionym         https:/… https://id.biodiv…
#>  6 Lycopodium lock… Lycopodium loc… basionym         https:/… https://id.biodiv…
#>  7 Aralia polaris   Aralia polaris… basionym         https:/… https://id.biodiv…
#>  8 Sium nodiflorum  Sium nodifloru… basionym         https:/… https://id.biodiv…
#>  9 Trachymene koch… Trachymene koc… basionym         https:/… https://id.biodiv…
#> 10 Pittosporum res… Pittosporum re… basionym         https:/… https://id.biodiv…
#> # ℹ 69,637 more rows
#> # ℹ 10 more variables: accepted_name_usage_ID <chr>, name_type <chr>,
#> #   taxon_rank <chr>, genus <chr>, stripped_canonical <chr>,
#> #   stripped_canonical2 <chr>, stripped_scientific <chr>, binomial <chr>,
#> #   trinomial <chr>, taxonomic_dataset <chr>
#> 
#> $`APNI names`
#> # A tibble: 15,379 × 13
#>    canonical_name        scientific_name scientific_name_ID name_type taxon_rank
#>    <chr>                 <chr>           <chr>              <chr>     <chr>     
#>  1 (Eucalyptus globulus… (Eucalyptus gl… https://id.biodiv… hybrid-f… species   
#>  2 (Eucalyptus globulus… (Eucalyptus gl… https://id.biodiv… hybrid-f… species   
#>  3 (Eucalyptus johnston… (Eucalyptus jo… https://id.biodiv… hybrid-f… species   
#>  4 (Eucalyptus johnston… (Eucalyptus jo… https://id.biodiv… hybrid-f… species   
#>  5 (Eucalyptus urnigera… (Eucalyptus ur… https://id.biodiv… hybrid-f… species   
#>  6 (Eucalyptus urnigera… (Eucalyptus ur… https://id.biodiv… hybrid-f… species   
#>  7 Abacopteris presliana Abacopteris pr… https://id.biodiv… scientif… species   
#>  8 Abarema clypearia f.… Abarema clypea… https://id.biodiv… scientif… form      
#>  9 Abarema pruinosa      Abarema pruino… https://id.biodiv… scientif… species   
#> 10 Abelia rupestris      Abelia rupestr… https://id.biodiv… scientif… species   
#> # ℹ 15,369 more rows
#> # ℹ 8 more variables: taxonomic_status <chr>, stripped_canonical <chr>,
#> #   stripped_canonical2 <chr>, stripped_scientific <chr>, binomial <chr>,
#> #   trinomial <chr>, genus <chr>, taxonomic_dataset <chr>
#> 
#> $genera_accepted
#> # A tibble: 3,094 × 11
#>    canonical_name accepted_name_usage accepted_name_usage_ID     scientific_name
#>    <chr>          <chr>               <chr>                      <chr>          
#>  1 Cycas          Cycas L.            https://id.biodiversity.o… Cycas L.       
#>  2 Bowenia        Bowenia Hook.f.     https://id.biodiversity.o… Bowenia Hook.f.
#>  3 Lepidozamia    Lepidozamia Regel   https://id.biodiversity.o… Lepidozamia Re…
#>  4 Macrozamia     Macrozamia Miq.     https://id.biodiversity.o… Macrozamia Miq.
#>  5 Equisetum      Equisetum L.        https://id.biodiversity.o… Equisetum L.   
#>  6 Isoetes        Isoetes L.          https://id.biodiversity.o… Isoetes L.     
#>  7 Huperzia       Huperzia Bernh.     https://id.biodiversity.o… Huperzia Bernh.
#>  8 Lycopodiella   Lycopodiella Holub  https://id.biodiversity.o… Lycopodiella H…
#>  9 Lycopodium     Lycopodium L.       https://id.biodiversity.o… Lycopodium L.  
#> 10 Phlegmariurus  Phlegmariurus Holub https://id.biodiversity.o… Phlegmariurus …
#> # ℹ 3,084 more rows
#> # ℹ 7 more variables: taxonomic_status <chr>, taxon_ID <chr>,
#> #   scientific_name_ID <chr>, name_type <chr>, taxon_rank <chr>, genus <chr>,
#> #   taxonomic_dataset <chr>
#> 
#> $genera_known
#> # A tibble: 2,852 × 11
#>    canonical_name   accepted_name_usage accepted_name_usage_ID   scientific_name
#>    <chr>            <chr>               <chr>                    <chr>          
#>  1 Encephallartes   Encephalartos Lehm. https://id.biodiversity… Encephallartes…
#>  2 Encephalartus    Encephalartos Lehm. https://id.biodiversity… Encephalartus …
#>  3 Encephalartos    Encephalartos Lehm. https://id.biodiversity… Encephalartos …
#>  4 Catakidozamia    Lepidozamia Regel   https://id.biodiversity… Catakidozamia …
#>  5 Zamia            Zamia L.            https://id.biodiversity… Zamia L.       
#>  6 Calamaria        Isoetes L.          https://id.biodiversity… Calamaria Boeh…
#>  7 Lateristachys    Lycopodiella Holub  https://id.biodiversity… Lateristachys …
#>  8 Lepidotis        Lycopodiella Holub  https://id.biodiversity… Lepidotis P.Be…
#>  9 Austrolycopodium Lycopodium L.       https://id.biodiversity… Austrolycopodi…
#> 10 Diphasium        Lycopodium L.       https://id.biodiversity… Diphasium C.Pr…
#> # ℹ 2,842 more rows
#> # ℹ 7 more variables: taxonomic_status <chr>, taxon_ID <chr>,
#> #   scientific_name_ID <chr>, name_type <chr>, taxon_rank <chr>, genus <chr>,
#> #   taxonomic_dataset <chr>
#> 
#> $genera_APNI
#> # A tibble: 1,430 × 8
#>    canonical_name scientific_name  taxonomic_status scientific_name_ID name_type
#>    <chr>          <chr>            <chr>            <chr>              <chr>    
#>  1 Beloperone     Beloperone Nees  unplaced         https://id.biodiv… scientif…
#>  2 Hemiadelphis   Hemiadelphis Ne… unplaced         https://id.biodiv… scientif…
#>  3 Leptosiphonium Leptosiphonium … unplaced         https://id.biodiv… scientif…
#>  4 Megaskepasma   Megaskepasma Li… unplaced         https://id.biodiv… scientif…
#>  5 Meyenia        Meyenia Nees     unplaced         https://id.biodiv… scientif…
#>  6 Simonachne     Simonachne E.J.… unplaced         https://id.biodiv… scientif…
#>  7 Thyrsacanthus  Thyrsacanthus N… unplaced         https://id.biodiv… scientif…
#>  8 Synarrhena     Synarrhena F.Mu… unplaced         https://id.biodiv… scientif…
#>  9 Allosorus      Allosorus Bernh. unplaced         https://id.biodiv… scientif…
#> 10 Austrogramme   Austrogramme E.… unplaced         https://id.biodiv… scientif…
#> # ℹ 1,420 more rows
#> # ℹ 3 more variables: taxon_rank <chr>, genus <chr>, taxonomic_dataset <chr>
#> 
#> $genera_all
#> # A tibble: 7,376 × 12
#>    canonical_name accepted_name_usage accepted_name_usage_ID     scientific_name
#>    <chr>          <chr>               <chr>                      <chr>          
#>  1 Cycas          Cycas L.            https://id.biodiversity.o… Cycas L.       
#>  2 Bowenia        Bowenia Hook.f.     https://id.biodiversity.o… Bowenia Hook.f.
#>  3 Lepidozamia    Lepidozamia Regel   https://id.biodiversity.o… Lepidozamia Re…
#>  4 Macrozamia     Macrozamia Miq.     https://id.biodiversity.o… Macrozamia Miq.
#>  5 Equisetum      Equisetum L.        https://id.biodiversity.o… Equisetum L.   
#>  6 Isoetes        Isoetes L.          https://id.biodiversity.o… Isoetes L.     
#>  7 Huperzia       Huperzia Bernh.     https://id.biodiversity.o… Huperzia Bernh.
#>  8 Lycopodiella   Lycopodiella Holub  https://id.biodiversity.o… Lycopodiella H…
#>  9 Lycopodium     Lycopodium L.       https://id.biodiversity.o… Lycopodium L.  
#> 10 Phlegmariurus  Phlegmariurus Holub https://id.biodiversity.o… Phlegmariurus …
#> # ℹ 7,366 more rows
#> # ℹ 8 more variables: taxonomic_status <chr>, taxon_ID <chr>,
#> #   scientific_name_ID <chr>, name_type <chr>, taxon_rank <chr>, genus <chr>,
#> #   taxonomic_dataset <chr>, cleaned_name <chr>
#> 
#> $family_accepted
#> # A tibble: 307 × 38
#>    taxon_ID                 name_type accepted_name_usage_ID accepted_name_usage
#>    <chr>                    <chr>     <chr>                  <chr>              
#>  1 https://id.biodiversity… scientif… https://id.biodiversi… Cycadaceae Pers.   
#>  2 https://id.biodiversity… scientif… https://id.biodiversi… Zamiaceae Horan.   
#>  3 https://id.biodiversity… scientif… https://id.biodiversi… Equisetaceae Michx…
#>  4 https://id.biodiversity… scientif… https://id.biodiversi… Isoetaceae Rchb.   
#>  5 https://id.biodiversity… scientif… https://id.biodiversi… Lycopodiaceae P.Be…
#>  6 https://id.biodiversity… scientif… https://id.biodiversi… Selaginellaceae Wi…
#>  7 https://id.biodiversity… scientif… https://id.biodiversi… Apiaceae Lindl.    
#>  8 https://id.biodiversity… scientif… https://id.biodiversi… Araliaceae Juss.   
#>  9 https://id.biodiversity… scientif… https://id.biodiversi… Griseliniaceae J.R…
#> 10 https://id.biodiversity… scientif… https://id.biodiversi… Myodocarpaceae Dow…
#> # ℹ 297 more rows
#> # ℹ 34 more variables: nomenclaturalStatus <chr>, nomIlleg <chr>,
#> #   nomInval <chr>, taxonomic_status <chr>, pro_parte <lgl>,
#> #   scientific_name <chr>, scientific_name_ID <chr>, canonical_name <chr>,
#> #   scientific_name_authorship <chr>, parentNameUsageID <chr>,
#> #   taxon_rank <chr>, taxon_rank_sort_order <dbl>, kingdom <chr>, class <chr>,
#> #   subclass <chr>, family <chr>, taxonConceptID <chr>, …
#>