Use APC and APNI to update taxonomy, replacing synonyms to current taxa where relevant

This function uses the Australia's Virtual Herbarium's taxonomic resources, specifically the Australian Plant Census (APC) and the Australian Plant Name Index (APNI), to update taxonomy of plant species, replacing any synonyms to their current accepted name.

Usage

update_taxonomy(
  aligned_names,
  output = NULL,
  resources = load_taxonomic_resources()
)

Arguments

aligned_names

A character vector of plant names to update. These names must be in the format of the scientific name, with genus and species, and may contain additional qualifiers such as subspecies or varieties. The names are case insensitive.

output

(optional) Name of the file where results are saved. The default is NULL and no file is created. If specified, the output will be saved in a CSV file with the given name.

resources

the taxonomic resources required to make the summary statistics. Loading this can be slow, so call load_taxonomic_resources separately to greatly speed this function up and pass the resources in.

Value

A tibble with updated taxonomy for the specified plant names. The tibble contains the following columns:

aligned_name: the input plant name.

source: the source of the updated taxonomic information (APC or APNI).

taxonIDClean: the unique identifier for the updated taxon.

taxonomicStatusClean: the taxonomic status of the updated taxon.

alternativeTaxonomicStatusClean: the alternative taxonomic status for the input name, if any.

acceptedNameUsageID: the unique identifier for the accepted name of the input name.

canonicalName: the accepted scientific name for the input name.

scientificNameAuthorship: the authorship information for the accepted name.

taxonRank: the taxonomic rank of the accepted name.

taxonomicStatus: the taxonomic status of the accepted name.

family: the family of the accepted name.

subclass: the subclass of the accepted name.

taxonDistribution: the distribution of the accepted name.

ccAttributionIRI: the Creative Commons Attribution International Rights URI of the accepted name.

See also

Examples

# Update taxonomy for two plant names and print the result
update_taxonomy(c("Eucalyptus pauciflora", "Acacia victoriae"))
#> Loading resources...
#> ...done
#> # A tibble: 2 × 14
#>   aligned_name   source taxonIDClean taxonomicStatusClean alternativeTaxonomic…¹
#>   <chr>          <chr>  <chr>        <chr>                <chr>                 
#> 1 Acacia victor… APC    https://id.… accepted             NA                    
#> 2 Eucalyptus pa… APC    https://id.… accepted             NA                    
#> # ℹ abbreviated name: ¹​alternativeTaxonomicStatusClean
#> # ℹ 9 more variables: acceptedNameUsageID <chr>, canonicalName <chr>,
#> #   scientificNameAuthorship <chr>, taxonRank <chr>, taxonomicStatus <chr>,
#> #   family <chr>, subclass <chr>, taxonDistribution <chr>,
#> #   ccAttributionIRI <chr>